+++ nodejs/aws-infra/experimental/index.ts
@@ -21,5 +21,5 @@ export * from "./containerDefinition";
 export * from "./image";
 export * from "./ec2";
 export * from "./fargate";
-export * from "./port";
+export * from "./loadBalancer";
 export * from "./taskDefinition";
